Fix cannot upload model consecutively bug
The current implementation will rely on the return of hash computing
backend function to determine whether or not upload the model. IF
the hash isn't in the csv file on the disk, that function will always
return `None`, which will not trigger the textbox onchange event, which
will not run the function that evaluates whether it needs to upload
the model or not.