Add a flag to monitor whetehr the user need to upgrade their plan
Use a gradio checkbox as a flag on the front end to monitor whether the user need
to upgrade to a newer plan.

If there is not enough inference points available, if will check the
checkbox, and there is an event listen on the checkbox to redirect users
to payment.