Change the preview loading logic to leverage browser cache
We have changed the preview loading logic to leverage the browser
cache. In the previous version, we would like to enable users to save
private previews, we add a timmestamp at the end of each image previews.
The prpblem is that every time the image url will change when refresh or
opened, and cache memachnism is totally ignored.

However, while more and more models are being added, we would like to
give this capability back. The users are still be able to add their
private model previews, but the ones that does not change will be cached
by browser.