Do not throw exceptions in thread pool executor
Throwing exceptions in thread pool executor will attach the exception
frame to the thread object, which will not be destroyed and gced. So
instead of throwing an exception, we will construct the return directly
in the thread worker function and return as string, which will make
exception frame go out of scope after the `try` and `except` clause.